requirements.txt
coderedcms==0.20.*
django-ckeditor==6.1.0